Move music library tests to their own file
